avoid sleeping during tests
